Add more information for how to use Model#hides
I am new to Nodal and would like to help with the documentation so please just let me know if there is a better or more appropriate place to submit this kind of update! :D

This particular change is because I spent quite a bit of time trying to figure out how to use hides() before I found the example in the Dev Logs release notes.